factor "doc-test-only" code into its own crate
