+++ src/sacp-doc-test/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "sacp-doc-test"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+sacp = { path = "../sacp" }
+serde.workspace = true
+futures.workspace = true
+serde_json.workspace = true
