+++ src/sacp/Cargo.toml
@@ -25,3 +25,4 @@ expect-test.workspace = true
 sacp-tokio = { version = "0.1.1", path = "../sacp-tokio" }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "io-util", "time"] }
 tokio-util.workspace = true
+sacp-doc-test = { path = "../sacp-doc-test" }
