+++ src/sacp/src/lib.rs
@@ -15,7 +15,3 @@ pub use agent_client_protocol_schema::*;
 pub use acp::*;
 pub use capabilities::*;
 pub use jsonrpc::*;
-
-// Re-export doc_test module for doc tests
-#[doc(hidden)]
-pub mod doc_test;
